Restructured project file contents to better organize project directory.
