+++ burstphoto/constants.h
@@ -1,0 +1,17 @@
+//
+//  constants.metal
+//  burstphoto
+//
+//  Created by Alex on 2023-07-03.
+//
+
+#include <metal_stdlib>
+using namespace metal;
+
+
+constant uint UINT16_MAX_VAL = 65535;
+constant float PI = 3.14159265358979323846f;
+constant half FLOAT16_ZERO_VAL = half(0);
+constant half FLOAT16_MIN_VAL = half(-65504);
+constant half FLOAT16_MAX_VAL = half(65504);
+constant half FLOAT16_05_VAL = half(0.5f);
