+++ burstphoto/merge/spatial.metal
@@ -1,0 +1,74 @@
+#include <metal_stdlib>
+using namespace metal;
+
+
+kernel void add_texture_weighted(texture2d<float, access::read> texture1 [[texture(0)]],
+                                 texture2d<float, access::read> texture2 [[texture(1)]],
+                                 texture2d<float, access::read> weight_texture [[texture(2)]],
+                                 texture2d<float, access::write> out_texture [[texture(3)]],
+                                 uint2 gid [[thread_position_in_grid]]) {
+    
+    float intensity1 = texture1.read(gid).r;
+    float intensity2 = texture2.read(gid).r;
+    float weight = weight_texture.read(gid).r;
+    
+    float out_intensity = weight * intensity2 + (1 - weight) * intensity1;
+    out_texture.write(out_intensity, gid);
+}
+
+kernel void color_difference(texture2d<float, access::read> texture1 [[texture(0)]],
+                             texture2d<float, access::read> texture2 [[texture(1)]],
+                             texture2d<float, access::write> out_texture [[texture(2)]],
+                             constant int& mosaic_pattern_width [[buffer(0)]],
+                             uint2 gid [[thread_position_in_grid]]) {
+    
+    float total_diff = 0;
+    int x0 = gid.x * mosaic_pattern_width;
+    int y0 = gid.y * mosaic_pattern_width;
+    
+    for (int dx = 0; dx < mosaic_pattern_width; dx++) {
+        for (int dy = 0; dy < mosaic_pattern_width; dy++) {
+            int x = x0 + dx;
+            int y = y0 + dy;
+            float i1 = texture1.read(uint2(x, y)).r;
+            float i2 = texture2.read(uint2(x, y)).r;
+            total_diff += abs(i1 - i2);
+        }
+    }
+    
+    out_texture.write(total_diff, gid);
+}
+
+kernel void compute_merge_weight(texture2d<float, access::read> texture_diff [[texture(0)]],
+                                 texture2d<float, access::write> weight_texture [[texture(1)]],
+                                 constant float* noise_sd_buffer [[buffer(0)]],
+                                 constant float& robustness [[buffer(1)]],
+                                 uint2 gid [[thread_position_in_grid]]) {
+    
+    // load args
+    float noise_sd = noise_sd_buffer[0];
+    
+    // load texture difference
+    float diff = texture_diff.read(gid).r;
+    
+    // compute the weight to assign to the comparison frame
+    // weight == 0 means that the aligned image is ignored
+    // weight == 1 means that the aligned image has full weight
+    float weight;
+    if (robustness == 0) {
+        // robustness == 0 means that robust merge is turned off
+        weight = 1;
+    } else {
+        // compare the difference to image noise
+        // as diff increases, the weight of the aligned image will continuously decrease from 1.0 to 0.0
+        // the two extreme cases are:
+        // diff == 0                   --> aligned image will have weight 1.0
+        // diff >= noise_sd/robustness --> aligned image will have weight 0.0
+        float max_diff = noise_sd / robustness;
+        weight =  1 - diff / max_diff;
+        weight = clamp(weight, 0.0, 1.0);
+    }
+    
+    // write weight
+    weight_texture.write(weight, gid);
+}
